code cleanup on milestones
