+++ src/components/Settings/SettingBody/MilestoneDetails/MilestoneTable/MilestoneTable.tsx
@@ -1,25 +1,23 @@
 import React, { useState } from 'react';
 
-import Box from '@mui/material/Box';
-import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
-import TableCell from '@mui/material/TableCell';
-import TableContainer from '@mui/material/TableContainer';
-import TableHead from '@mui/material/TableHead';
-import TablePagination from '@mui/material/TablePagination';
-import TableRow from '@mui/material/TableRow';
-import TableSortLabel from '@mui/material/TableSortLabel';
+import {
+  Grid,
+  FormControlLabel,
+  Paper,
+  Switch,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TablePagination,
+  TableRow,
+} from '@mui/material';
+import MilestoneRow from '@components/Settings/SettingBody/MilestoneDetails/MilestoneTable/MilestoneRow/MilestoneRow';
+import MilestoneTableHeader from '@components/Settings/SettingBody/MilestoneDetails/MilestoneTable/MilestoneTableHeader/MilestoneTableHeader';
 
-import Paper from '@mui/material/Paper';
-
-import FormControlLabel from '@mui/material/FormControlLabel';
-import Switch from '@mui/material/Switch';
-
-import { visuallyHidden } from '@mui/utils';
-import { Grid } from '@mui/material';
 import { MilestoneData } from '@/modules/Milestones/Types';
 import DateTime from '@/modules/DateTime/DateTime';
-import MilestoneRow from './MilestoneRow/MilestoneRow';
+import { Order } from '@/consts/constants';
 
 type Props = {
   milestones: MilestoneData[];
@@ -35,8 +33,6 @@ function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
   }
   return 0;
 }
-
-type Order = 'asc' | 'desc';
 
 function getComparator<Key extends keyof MilestoneData>(
   order: Order,
@@ -57,84 +53,6 @@ function stableSort<T>(array: readonly T[], comparator: (a: T, b: T) => number) 
     return a[1] - b[1];
   });
   return stabilizedThis.map(el => el[0]);
-}
-
-interface HeadCell {
-  disablePadding: boolean;
-  id: keyof MilestoneData;
-  label: string;
-  numeric: boolean;
-}
-
-const headCells: readonly HeadCell[] = [
-  {
-    id: 'weekBeginning',
-    numeric: false,
-    disablePadding: false,
-    label: 'Date',
-  },
-  {
-    id: 'week',
-    numeric: true,
-    disablePadding: false,
-    label: 'Week',
-  },
-];
-
-interface EnhancedTableProps {
-  onRequestSort: (event: React.MouseEvent<unknown>, property: keyof MilestoneData) => void;
-  order: Order;
-  orderBy: string;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-  const { order, orderBy, onRequestSort } = props;
-  const createSortHandler = (property: keyof MilestoneData) => (event: React.MouseEvent<unknown>) => {
-    onRequestSort(event, property);
-  };
-
-  return (
-    <TableHead className='bg-[#4D4D4D]'>
-      <TableRow>
-        {headCells.map(headCell => (
-          <TableCell
-            className='text-bgWhite'
-            key={headCell.id}
-            align={headCell.numeric ? 'right' : 'left'}
-            padding={headCell.disablePadding ? 'none' : 'normal'}
-            sortDirection={orderBy === headCell.id ? order : false}
-          >
-            <TableSortLabel
-              active={orderBy === headCell.id}
-              direction={orderBy === headCell.id ? order : 'asc'}
-              onClick={createSortHandler(headCell.id)}
-              sx={{
-                '&.MuiTableSortLabel-root': {
-                  color: 'white',
-                },
-                '&.MuiTableSortLabel-root:hover': {
-                  color: '#E33939',
-                },
-                '&.Mui-active': {
-                  color: 'white',
-                },
-                '& .MuiTableSortLabel-icon': {
-                  color: '#E33939 !important',
-                },
-              }}
-            >
-              {headCell.label}
-              {orderBy === headCell.id ? (
-                <Box component='span' sx={visuallyHidden}>
-                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
-                </Box>
-              ) : null}
-            </TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
 }
 
 const MilestoneTable = (props: Props) => {
@@ -173,7 +91,7 @@ const MilestoneTable = (props: Props) => {
         <Paper className='drop-shadow-2xl' sx={{ width: '100%', mb: 2 }}>
           <TableContainer>
             <Table sx={{ minWidth: 750 }} size={dense ? 'small' : 'medium'}>
-              <EnhancedTableHead order={order} orderBy={orderBy} onRequestSort={handleRequestSort} />
+              <MilestoneTableHeader order={order} orderBy={orderBy} onRequestSort={handleRequestSort} />
               <TableBody>
                 {/* if you don't need to support IE11, you can replace the `stableSort` call with:
               rows.slice().sort(getComparator(order, orderBy)) */}
